backend/requirements.txt
Django==6.0.4
channels==4.3.2
channels-redis==4.3.0
chess==1.11.2
daphne==4.2.1
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.3.2
